Split mesh caches to have a Host and Device implementation
