+++ src/mesh/MeshColumns.cc
@@ -21,14 +21,14 @@ namespace AmanziMesh {
 // Constructor that guesses how to infer columnar structure.
 //
 void
-MeshColumns::initialize(const MeshCache<MemSpace_kind::HOST>& mesh)
+MeshColumns::initialize(const MeshHost& mesh)
 {
   AMANZI_ASSERT(mesh.getSpaceDimension() == 3);
   AMANZI_ASSERT(mesh.getManifoldDimension() == 3);
 
   // loop over all boundary faces and look for those whose normal z component
   // is not zero and whose opposing face is below them.
-  MeshCache<MemSpace_kind::HOST>::Entity_ID_View surface_faces("surface_faces",
+  MeshHost::Entity_ID_View surface_faces("surface_faces",
                                                                mesh.getBoundaryFaces().size());
   int sf = 0;
   for (const Entity_ID f : mesh.getBoundaryFaces()) {
@@ -45,7 +45,7 @@ MeshColumns::initialize(const MeshCache<MemSpace_kind::HOST>& mesh)
 // Constructor that infers columnar structure from a mesh set.
 //
 void
-MeshColumns::initialize(const MeshCache<MemSpace_kind::HOST>& mesh,
+MeshColumns::initialize(const MeshHost& mesh,
                         const std::vector<std::string>& regions)
 {
   AMANZI_ASSERT(mesh.getSpaceDimension() == 3);
@@ -63,7 +63,7 @@ MeshColumns::initialize(const MeshCache<MemSpace_kind::HOST>& mesh,
     for (Entity_ID f : r_faces)
       vsurface_faces.insert(std::upper_bound(vsurface_faces.begin(), vsurface_faces.end(), f), f);
   }
-  MeshCache<MemSpace_kind::HOST>::Entity_ID_View surface_faces;
+  MeshHost::Entity_ID_View surface_faces;
   vectorToView(surface_faces, vsurface_faces);
 
   // build columns for each face
@@ -72,8 +72,8 @@ MeshColumns::initialize(const MeshCache<MemSpace_kind::HOST>& mesh,
 
 
 void
-MeshColumns::initialize(const MeshCache<MemSpace_kind::HOST>& mesh,
-                        const MeshCache<MemSpace_kind::HOST>::Entity_ID_View& surface_faces)
+MeshColumns::initialize(const MeshHost& mesh,
+                        const MeshHost::Entity_ID_View& surface_faces)
 {
   // figure out the correct size
   // Note, this is done to make life easier for Kokkos
@@ -114,7 +114,7 @@ MeshColumns::initialize(const MeshCache<MemSpace_kind::HOST>& mesh,
 // Actually does the work of building the column, starting at f.
 //
 void
-MeshColumns::buildColumn_(const MeshCache<MemSpace_kind::HOST>& mesh, Entity_ID f, int col)
+MeshColumns::buildColumn_(const MeshHost& mesh, Entity_ID f, int col)
 {
   Parallel_kind ptype = mesh.getParallelType(Entity_kind::FACE, f);
 
@@ -158,7 +158,7 @@ MeshColumns::buildColumn_(const MeshCache<MemSpace_kind::HOST>& mesh, Entity_ID 
 namespace Impl {
 
 int
-orientFace(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID f, const Entity_ID c)
+orientFace(const MeshHost& mesh, const Entity_ID f, const Entity_ID c)
 {
   auto normal = mesh.getFaceNormal(f, c);
   normal /= AmanziGeometry::norm(normal);
@@ -176,7 +176,7 @@ orientFace(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID f, const 
 // Assumes this is well posed... e.g. that there is only one face of c that
 // shares no nodes with f.   Returns -1 if this is not possible.
 Entity_ID
-findDownFace(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID c)
+findDownFace(const MeshHost& mesh, const Entity_ID c)
 {
   auto cfaces = mesh.getCellFaces(c);
   // find the one that is down
@@ -190,7 +190,7 @@ findDownFace(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID c)
 // Finds the cell in face cells that is not c, returning -1 if not possible.
 //
 Entity_ID
-findOpposingCell(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID c, const Entity_ID f)
+findOpposingCell(const MeshHost& mesh, const Entity_ID c, const Entity_ID f)
 {
   auto fcells = mesh.getFaceCells(f);
   if (fcells.size() == 1)
@@ -205,7 +205,7 @@ findOpposingCell(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID c, 
 // Helper function for counting column size
 //
 std::size_t
-countCellsInColumn(const MeshCache<MemSpace_kind::HOST>& mesh, Entity_ID f)
+countCellsInColumn(const MeshHost& mesh, Entity_ID f)
 {
   std::size_t count = 0;
   Entity_ID c = mesh.getFaceCells(f)[0]; // guaranteed size 1
