+++ src/mesh/MeshColumns.hh
@@ -18,8 +18,9 @@
 namespace Amanzi {
 namespace AmanziMesh {
 
-template <MemSpace_kind>
-struct MeshCache;
+struct MeshCacheHost;
+struct MeshCacheDevice;
+
 
 struct MeshColumns {
  public:
@@ -34,7 +35,7 @@ struct MeshColumns {
   //
   // Note these MUST be constructed on HOST caches.
   //
-  void initialize(const MeshCache<MemSpace_kind::HOST>& mesh);
+  void initialize(const MeshCacheHost& mesh);
 
   //
   // Constructor that infers columnar structure from a mesh set.
@@ -42,14 +43,14 @@ struct MeshColumns {
   // Note these MUST be constructed on HOST caches.
   //
   void
-  initialize(const MeshCache<MemSpace_kind::HOST>& mesh, const std::vector<std::string>& regions);
+  initialize(const MeshCacheHost& mesh, const std::vector<std::string>& regions);
 
   //
   // Constructor that sets top faces directly
   //
   // Note these MUST be constructed on HOST caches.
   //
-  void initialize(const MeshCache<MemSpace_kind::HOST>& mesh,
+  void initialize(const MeshCacheHost& mesh,
                   const View_type<Entity_ID, MemSpace_kind::HOST>& surface_faces);
 
   int num_columns_owned;
@@ -88,7 +89,7 @@ struct MeshColumns {
   }
 
  private:
-  void buildColumn_(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID f, int col);
+  void buildColumn_(const MeshCacheHost& mesh, const Entity_ID f, int col);
 
  public:
   // a vector of columns, each a list of cells in the column, ordered from top
@@ -108,7 +109,7 @@ namespace Impl {
 // Returns 1 if the up face, -1 if the down face, 0 if the side face.
 //
 int
-orientFace(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID f, const Entity_ID c);
+orientFace(const MeshCacheHost& mesh, const Entity_ID f, const Entity_ID c);
 
 //
 // Helper function to find the face of c pointing downward.
@@ -116,19 +117,19 @@ orientFace(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID f, const 
 // Assumes this is well posed... e.g. that there is only one face of c that
 // points down.   Returns -1 if none are found.
 Entity_ID
-findDownFace(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID c);
+findDownFace(const MeshCacheHost& mesh, const Entity_ID c);
 
 //
 // Finds the cell in face cells that is not c, returning -1 if not possible.
 //
 Entity_ID
-findOpposingCell(const MeshCache<MemSpace_kind::HOST>& mesh, const Entity_ID c, const Entity_ID f);
+findOpposingCell(const MeshCacheHost& mesh, const Entity_ID c, const Entity_ID f);
 
 //
 // Helper function for counting column size
 //
 std::size_t
-countCellsInColumn(const MeshCache<MemSpace_kind::HOST>& mesh, Entity_ID f);
+countCellsInColumn(const MeshCacheHost& mesh, Entity_ID f);
 
 } // namespace Impl
