+++ src/mesh/mesh_extracted/MeshFrameworkColumn.cc
@@ -70,7 +70,7 @@ MeshFrameworkColumn::computeSpecialNodeCoordinates_()
   // error will be zero.
 
   // Create a cached object, so that we can use columns.
-  MeshCache<MemSpace_kind::HOST> col3D_mesh(
+  MeshHost col3D_mesh(
     col3D_mesh_, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null);
   col3D_mesh.buildColumns();
