+++ src/mesh/mesh_extracted/test/mesh_extracted_fracture.cc
@@ -97,7 +97,7 @@ RunTest(const std::string regname, int* cells, int* edges, const std::vector<int
     // extract fractures mesh
     try {
       auto mesh3D_cache = Teuchos::rcp(
-        new Mesh(mesh3D, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+        new MeshHost(mesh3D, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
       cacheAll(*mesh3D_cache);
 
       RCP<MeshFramework> mesh = Teuchos::rcp(new MeshExtractedManifold(
@@ -125,12 +125,14 @@ RunTest(const std::string regname, int* cells, int* edges, const std::vector<int
 
       auto plist_tmp = Teuchos::rcp(new Teuchos::ParameterList);
       plist_tmp->set<bool>("natural map ordering", true);
-      RCP<Mesh> mesh_cache =
-        Teuchos::rcp(new Mesh(mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), plist_tmp));
+      RCP<MeshHost> mesh_cache =
+        Teuchos::rcp(new MeshHost(mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), plist_tmp));
       cacheAll(*mesh_cache);
 
+      auto mesh_cache_device = onMemDevice(mesh_cache); 
+
       // verify mesh
-      MeshAudit audit(mesh_cache);
+      MeshAudit audit(mesh_cache_device);
       int ok = audit.Verify();
       CHECK(ok == 0);
