+++ src/mesh/mesh_logical/test/demo_mesh.cc
@@ -359,7 +359,7 @@ demoMeshLogicalYEmbedded()
   auto comm = Amanzi::getDefaultComm();
 
   Teuchos::RCP<MeshFramework> m_log_fw = demoMeshLogicalYManual();
-  Teuchos::RCP<Mesh> m_log = Teuchos::rcp(new Mesh(
+  Teuchos::RCP<MeshHost> m_log = Teuchos::rcp(new MeshHost(
     m_log_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
 
   Point X0(-1.75, -1.75, -3.);
@@ -379,7 +379,7 @@ demoMeshLogicalYEmbedded()
                                                                m_log->getGeometricModel(),
                                                                Teuchos::null));
   auto m_bg = Teuchos::rcp(
-    new Mesh(m_bg_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshAlgorithms()), Teuchos::null));
+    new MeshHost(m_bg_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshAlgorithms()), Teuchos::null));
 
   // make the new connections, 1 per logical cell
   int ncells_log = m_log->getNumEntities(Entity_kind::CELL, Parallel_kind::ALL);
