+++ src/mesh/mesh_logical/test/mesh_logical_geometry.cc
@@ -316,7 +316,7 @@ TEST(MESH_LOGICAL_SEGMENT_REGULAR_MANUAL)
             << "-----------------------------------------------------" << std::endl;
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalSegmentRegularManual();
   std::cout << "Before Cache" << std::endl;
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
   std::cout << "After Cache" << std::endl;
   test_segment_regular(mesh, false);
@@ -329,7 +329,7 @@ TEST(MESH_LOGICAL_SEGMENT_REGULAR_XML)
             << "TEST: MeshLogical single segment, xml construction" << std::endl
             << "-----------------------------------------------------" << std::endl;
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalFromXML("regular");
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
   test_segment_regular(mesh, false);
 }
@@ -342,7 +342,7 @@ TEST(MESH_LOGICAL_SEGMENT_IRREGULAR_WITH_SETS)
             << "TEST: MeshLogical single segment, deformed" << std::endl
             << "-----------------------------------------------------" << std::endl;
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalSegmentIrregularManual();
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
   test_segment_irregular(mesh, true);
 }
@@ -356,7 +356,7 @@ TEST(MESH_LOGICAL_2Y_XML_WITH_SETS)
             << "-----------------------------------------------------" << std::endl;
 
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalFromXML("logical mesh 2Y");
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
   test_2Y(mesh, true);
 }
@@ -369,7 +369,7 @@ TEST(MESH_LOGICAL_Y)
             << "TEST: MeshLogical Y, manual construction" << std::endl
             << "-----------------------------------------------------" << std::endl;
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalYManual();
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
   test_Y(mesh, true);
 }
@@ -382,7 +382,7 @@ TEST(MESH_LOGICAL_Y_XML_WITH_SETS)
             << "TEST: MeshLogical Y, from XML" << std::endl
             << "-----------------------------------------------------" << std::endl;
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalFromXML("logical mesh Y");
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
   test_Y(mesh, true);
 }
@@ -395,7 +395,7 @@ TEST(MESH_EMBEDDED_Y)
             << "TEST: MeshLogical Y, embedded in background mesh" << std::endl
             << "-----------------------------------------------------" << std::endl;
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalYEmbedded();
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshEmbeddedLogicalAlgorithms()), Teuchos::null));
   Amanzi::AmanziMesh::MeshLogicalAudit audit(mesh, std::cout);
   CHECK(!audit.Verify());
@@ -409,7 +409,7 @@ TEST(MESH_SUBGRID_VARIABLE_TAU)
             << "TEST: subgrid mesh in travel time space" << std::endl
             << "-----------------------------------------------------" << std::endl;
   Teuchos::RCP<MeshFramework> mesh_fw = Amanzi::Testing::demoMeshLogicalFromXML("subgrid mesh");
-  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshCache<MemSpace_kind::HOST>(
+  auto mesh = Teuchos::rcp(new Amanzi::AmanziMesh::MeshHost(
     mesh_fw, Teuchos::rcp(new Amanzi::AmanziMesh::MeshLogicalAlgorithms()), Teuchos::null));
   Amanzi::AmanziMesh::MeshLogicalAudit audit(mesh, std::cout);
   CHECK(!audit.Verify());
