+++ src/mesh/test/geometry_harnesses.hh
@@ -10,6 +10,7 @@
 #pragma once
 
 #include "MeshFrameworkTraits.hh"
+#include "MeshAudit_impl.hh"
 
 using namespace Amanzi;
 using namespace Amanzi::AmanziMesh;
@@ -85,7 +86,7 @@ testMeshGeometry_(const Teuchos::RCP<Mesh_type>& mesh,
 
                            // Search for a cell with the same centroid in the expected list of centroids
                            int j = 0;
-                           for (; j != ncells; ++j) {
+                           for (; j != exp_cell_centroids.size(); ++j) {
                              auto diff = exp_cell_centroids[j] - centroid;
                              if (AmanziGeometry::norm(diff) < 1.0e-10) {
                                test1(c) = 0;
@@ -139,7 +140,7 @@ testMeshGeometry_(const Teuchos::RCP<Mesh_type>& mesh,
                            AmanziGeometry::Point centroid = m.getFaceCentroid(f);
 
                            int j = 0;
-                           for (; j < nfaces; ++j) {
+                           for (; j < exp_face_centroids.size(); ++j) {
                              auto diff = exp_face_centroids[j] - centroid;
                              if (AmanziGeometry::norm(diff) < 1.0e-10) {
                                test1(f) = 0;
@@ -205,7 +206,7 @@ testMeshGeometry_(const Teuchos::RCP<Mesh_type>& mesh,
 
                            AmanziGeometry::Point centroid = m.getNodeCoordinate(n);
                            int j = 0;
-                           for (; j < nnodes; ++j) {
+                           for (; j < exp_node_coordinates.size(); ++j) {
                              auto diff = exp_node_coordinates[j] - centroid;
                              if (AmanziGeometry::norm(diff) < 1.0e-10) {
                                test1(n) = 0;
@@ -458,20 +459,24 @@ testExteriorMapsUnitBox(const Teuchos::RCP<Mesh_type>& mesh, int nx, int ny, int
   auto& faces = mesh->getMap(AmanziMesh::Entity_kind::FACE, true);
   auto bface_ids = mesh->getBoundaryFaces();
 
-  for (int j = 0; j != bfaces->getLocalNumElements(); ++j) {
-    auto bf = faces->getLocalElement(bfaces->getGlobalElement(j));
-    CHECK_EQUAL(bface_ids[j], bf);
-    auto f_centroid = mesh->getFaceCentroid(bf);
-    bool found = false;
-    for (int i = 0; i != mesh->getManifoldDimension(); ++i) {
-      if (std::abs(f_centroid[i]) < 1e-10 || std::abs(f_centroid[i] - 1) < 1e-10) {
-        found = true;
-        break;
+  Kokkos::parallel_for("for: testExteriorMapsUnitBox", bfaces->getLocalNumElements(), KOKKOS_LAMBDA(const int j) {
+      auto bf = faces->getLocalElement(bfaces->getGlobalElement(j));
+      //CHECK_EQUAL(bface_ids[j], bf);
+      auto f_centroid = mesh->getFaceCentroid(bf);
+      bool found = false;
+      for (int i = 0; i != mesh->getManifoldDimension(); ++i) {
+        if (std::abs(f_centroid[i]) < 1e-10 || std::abs(f_centroid[i] - 1) < 1e-10) {
+          found = true;
+          break;
+        }
       }
-    }
-    if (!found) { std::cout << "not found: " << bf << " at " << f_centroid << std::endl; }
-    CHECK(found);
-  }
+      
+      if (!found) {
+        printf("not found: %d at %g,%g,%g \n", bf, f_centroid[0], f_centroid[1], f_centroid[2]);
+        assert(found);
+      }
+      //CHECK(found);
+    }); 
 
   // check nodes are on the boundary
   //
@@ -488,9 +493,8 @@ testExteriorMapsUnitBox(const Teuchos::RCP<Mesh_type>& mesh, int nx, int ny, int
 
   auto& bnodes = mesh->getMap(AmanziMesh::Entity_kind::BOUNDARY_NODE, true);
   auto& nodes = mesh->getMap(AmanziMesh::Entity_kind::NODE, true);
-  for (int j = 0; j != bnodes->getLocalNumElements(); ++j) {
-    std::cout << " bnode " << j << " GID " << bnodes->getGlobalElement(j) << " LID "
-              << nodes->getLocalElement(bnodes->getGlobalElement(j)) << std::endl;
+  Kokkos::parallel_for("for: testExteriorMapUnitBox", bnodes->getLocalNumElements(), KOKKOS_LAMBDA(const int j){
+      printf(" bnode %d GID %d LID %d\n",j,bnodes->getGlobalElement(j),nodes->getLocalElement(bnodes->getGlobalElement(j))); 
 
     auto bn = nodes->getLocalElement(bnodes->getGlobalElement(j));
     AmanziGeometry::Point nc;
@@ -500,17 +504,18 @@ testExteriorMapsUnitBox(const Teuchos::RCP<Mesh_type>& mesh, int nx, int ny, int
         std::abs(nc[1] - 1) < 1e-10 || std::abs(nc[2]) < 1e-10 || std::abs(nc[2] - 1) < 1e-10) {
       found = true;
     }
-    CHECK(found);
-  }
+    assert(found); 
+    //CHECK(found);
+    }); 
 }
 
 
 //
 // Test a columnar system
 //
-template <MemSpace_kind MEM>
+template <class MeshType>
 inline void
-testColumnsUniformDz(const MeshCache<MEM>& mesh, double dz)
+testColumnsUniformDz(const MeshType& mesh, double dz)
 {
   // tests the columnar structure of cells
   int n_columns = mesh.columns.num_columns_all;
@@ -522,7 +527,7 @@ testColumnsUniformDz(const MeshCache<MEM>& mesh, double dz)
   bool owned = true;
 
   for (int col = 0; col != n_columns; ++col) {
-    const auto& cells = mesh.columns.cells_.template getRowUnmanaged<MEM>(col);
+    const auto& cells = mesh.columns.cells_.template getRowUnmanaged<MeshType::MEM>(col);
 
     // check all owned cells first, then all ghosted
     if (owned) {
@@ -537,7 +542,7 @@ testColumnsUniformDz(const MeshCache<MEM>& mesh, double dz)
     }
 
     // check geometry
-    const auto& faces = mesh.columns.faces_.template getRowUnmanaged<MEM>(col);
+    const auto& faces = mesh.columns.faces_.template getRowUnmanaged<MeshType::MEM>(col);
     CHECK(faces.size() == (cells.size() + 1));
     for (int i = 0; i != cells.size(); ++i) {
       Entity_ID c = cells[i];
