+++ src/mesh/test/mesh_fracture_extraction.cc
@@ -54,15 +54,15 @@ TEST(MESH_FRACTURE_EXTRACTION_GENERATED)
     fac.set_preference({ frm });
     // Make cache of current mesh
     auto parent_mesh_cache = Teuchos::rcp(
-      new Mesh(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+      new MeshHost(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
     cacheAll(*parent_mesh_cache);
 
     auto mesh = fac.create(parent_mesh_cache, setnames, AmanziMesh::Entity_kind::FACE, false);
     cacheAll(*mesh);
-
+    
     // test the surface mesh as a fracture mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
 
     // check we found the right number of things
     int ncells =
@@ -125,15 +125,15 @@ TEST(MESH_FRACTURE_EXTRACTION_GENERATED_EXTRACTED_MANIFOLD)
     fac.set_preference({ frm });
     // Make cache of current mesh
     auto parent_mesh_cache = Teuchos::rcp(
-      new Mesh(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+      new MeshHost(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
     cacheAll(*parent_mesh_cache);
 
     auto mesh = fac.create(parent_mesh_cache, setnames, AmanziMesh::Entity_kind::FACE, false);
     cacheAll(*mesh);
-
+    
     // test the surface mesh as a fracture mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
 
     // check we found the right number of things
     int ncells =
@@ -178,15 +178,15 @@ TEST(MESH_FRACTURE_EXTRACTION_EXO)
     MeshFactory fac(comm, gm);
     fac.set_preference({ frm });
     auto parent_mesh_cache = Teuchos::rcp(
-      new Mesh(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+      new MeshHost(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
     cacheAll(*parent_mesh_cache);
     auto mesh =
-      fac.create(parent_mesh_cache, { "fractures-two" }, AmanziMesh::Entity_kind::FACE, false);
+      fac.create(parent_mesh_cache, std::vector<std::string>{ "fractures-two" }, AmanziMesh::Entity_kind::FACE, false);
     cacheAll(*mesh);
 
     // test the surface mesh as a fracture mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
 
     // check we found the right number of things
     int ncells =
@@ -248,15 +248,15 @@ TEST(MESH_FRACTURE_EXTRACTION_EXO_MANIFOLD)
     MeshFactory fac(comm, gm, fac_plist);
     fac.set_preference({ frm });
     auto parent_mesh_cache = Teuchos::rcp(
-      new Mesh(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+      new MeshHost(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
     cacheAll(*parent_mesh_cache);
 
     auto mesh = fac.create(parent_mesh_cache, setnames, AmanziMesh::Entity_kind::FACE, false);
     cacheAll(*mesh);
-
+    
     // test the surface mesh as a fracture mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
 
     // check we found the right number of things
     int ncells =
