+++ src/mesh/test/mesh_sets.cc
@@ -100,7 +100,7 @@ TEST(MESH_SETS_3CUBE_EXO)
 //               << "Testing 3D Box 3x3x3 Par with " << AmanziMesh::to_string(frm) << std::endl
 //               << "------------------------------------------------" << std::endl;
 //     auto mesh = createUnstructured(Preference{frm}, "test/hex_3x3x3_sets.par", comm, gm);
-//     auto mesh_host = Teuchos::rcpFromRef(onMemSpace<MemSpace_kind::HOST>(*mesh));
+//     auto mesh_host = Teuchos::rcpFromRef(onMemHost(*mesh));
 //     testHexMeshSets3x3x3(mesh_host, true, frm);
 //   }
 // }
