+++ src/mesh/test/mesh_surface_extraction.cc
@@ -72,14 +72,15 @@ TEST(MESH_SURFACE_EXTRACTION_GENERATED)
       fac.create(parent_mesh, top_faces, AmanziMesh::Entity_kind::FACE, true);
 
     // make a MeshCache
-    auto mesh = Teuchos::rcp(new Mesh(
+    auto mesh = Teuchos::rcp(new MeshHost(
       surface_framework_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+    // Get the host mesh
     mesh->setParentMesh(parent_mesh);
     cacheAll(*mesh);
 
     // test the surface mesh as a 3x3 quad mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
     // -- geometry
     testGeometryQuad(mesh, 3, 3);
     // -- exterior maps
@@ -134,14 +135,14 @@ TEST(MESH_SURFACE_EXTRACTION_EXO)
       fac.create(parent_mesh, top_faces, AmanziMesh::Entity_kind::FACE, true);
 
     // make a MeshCache
-    auto mesh = Teuchos::rcp(new Mesh(
+    auto mesh = Teuchos::rcp(new MeshHost(
       surface_framework_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
     mesh->setParentMesh(parent_mesh);
     cacheAll(*mesh);
 
     // test the surface mesh as a 3x3 quad mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
     // -- geometry
     testGeometryQuad(mesh, 3, 3);
     // -- exterior maps
@@ -201,16 +202,15 @@ TEST(MESH_SURFACE_EXTRACTION_GENERATED_EXTRACTED_MANIFOLD)
 
     // cache all, even edges here because Manifold needs them
     auto parent_mesh_cache = Teuchos::rcp(
-      new Mesh(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+      new MeshHost(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
     cacheAll(*parent_mesh_cache);
 
-    auto mesh =
-      fac.create(parent_mesh_cache, { "Top Face Plane" }, AmanziMesh::Entity_kind::FACE, true);
+    auto mesh = fac.create(parent_mesh_cache, std::vector<std::string>{ "Top Face Plane" }, AmanziMesh::Entity_kind::FACE, true);
     cacheAll(*mesh);
 
     // test the surface mesh as a 3x3 quad mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
     // -- geometry
     testGeometryQuad(mesh, 3, 3);
 
@@ -272,16 +272,16 @@ TEST(MESH_SURFACE_EXTRACTION_EXO_EXTRACTED_MANIFOLD)
     MeshFactory fac(comm, gm, fac_plist);
     fac.set_preference({ frm });
     auto parent_mesh_cache = Teuchos::rcp(
-      new Mesh(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
+      new MeshHost(parent_mesh, Teuchos::rcp(new AmanziMesh::MeshAlgorithms()), Teuchos::null));
     cacheAll(*parent_mesh_cache);
 
     auto mesh =
-      fac.create(parent_mesh_cache, { "Top Face Plane" }, AmanziMesh::Entity_kind::FACE, true);
+      fac.create(parent_mesh_cache, std::vector<std::string>{ "Top Face Plane" }, AmanziMesh::Entity_kind::FACE, true);
     cacheAll(*mesh);
 
     // test the surface mesh as a 3x3 quad mesh
     // -- mesh audit
-    testMeshAudit<MeshAudit, Mesh>(mesh);
+    testMeshAudit<MeshAuditHost, MeshHost>(mesh);
     // -- geometry
     testGeometryQuad(mesh, 3, 3);
     // -- exterior maps
